plus de norm

+++ asm/srcs/main.c
@@ -6,13 +6,13 @@
 /*   By: lubrun <lubrun@student.le-101.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/16 21:46:31 by lubrun            #+#    #+#             */
-/*   Updated: 2020/02/25 20:44:22 by lubrun           ###   ########lyon.fr   */
+/*   Updated: 2020/02/28 20:43:50 by lubrun           ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "asm.h"
 
-static int          is_asm_file(char *name)
+static int  is_asm_file(char *name)
 {
     int index;
 
@@ -24,80 +24,61 @@ static int          is_asm_file(char *name)
     return (1);
 }
 
-static t_option     parse_option(int ac, char **av)
-{
-    int         index;
-
-    index = 1;
-    while (index < ac)
-    {
-        if (av[index][0] != '-')
-            return ((t_option){0, -1, -1});
-        if (av[index][1] == 'a')
-            return ((t_option){1, 1, 0});
-        index++;
-    }
-    return ((t_option){0, -1, -1});
-}
-
-static t_option     print_error(int code, char *name)
+static int  print_error(int code, char *name)
 {
     if (code == 0)
-        write(1, "syntax error: ./asm [file.s][file1.s]...\noption:\n\t-a\tFor verbose\n", 66);
+        write(1, "syntax error: ./asm [file.s][file1.s]...\n", 41);
     else if (code == 1)
         write(1, "Can't write output file\n", 24);
-	unlink(name);
-    return ((t_option){0, -1, 1});
+    if (name)
+        unlink(name);
+    return (0);
 }
 
-static t_option     parse_arg(int ac, char **av)
+static int  parse_arg(int ac, char **av)
 {
-    t_option option;
-
-    if (ac <= 2)
+    if (ac == 1)
+        return (print_error(0, NULL));
+    if (ac == 2)
     {
-        if (is_asm_file(av[1]))
-            return ((t_option){0, 0, 0});
-        if (av[1] && ft_strcmp(av[1], "-help"))
+        if (av[1] && !is_asm_file(av[1]))
         {
-            write(1, "Syntax: ./asm -help\n", 21);
-            return ((t_option){-1, -1, 1});
+            write(1, "Only .s file accepted\n", 22);
+            return (0);
         }
     }
-    if ((option = parse_option(ac, av)).error)
-        return ((t_option){-1, -1, 1});
-    return (option);
+    return (1);
 }
 
-static int      parse_file(int ac, char **av, t_option opt)
+static int  parse_file(int ac, char **av)
 {
     t_file	*file;
     int		index;
 
-    index = opt.count + 1;
+    index = 1;
     while (index < ac)
     {
 		if (!(file = ft_memalloc(sizeof(t_file))))
 			return (0);
 	    ft_bzero(file, sizeof(t_file));
-		if (!is_asm_file(av[index]) || (file->fd = open(av[index], O_RDONLY)) == -1 ||
+		if (!is_asm_file(av[index]))
+            return (print_error(2, NULL));
+        if ((file->fd = open(av[index], O_RDONLY)) == -1 ||
 			!(file->file_name = ft_strdup(av[index])))
-            return (print_error(1, get_output_name(av[index])).count);
+            return (print_error(1, get_output_name(av[index])));
         if (!parse_header(file) || !parse_op(file) ||
 			!write_file(file))
-            return (print_error(file->error, get_output_name(av[index])).count);
+            return (print_error(file->error, get_output_name(av[index])));
 		index++;
         free_file(file);
     }
     return (1);
 }
 
-int              main(int ac, char **av)
+int         main(int ac, char **av)
 {
-    t_option    opt;
-
-    if ((opt = parse_arg(ac, av)).error ||
-        !parse_file(ac, av, opt))
-        return (opt.error);
+    if (!parse_arg(ac, av) ||
+        !parse_file(ac, av))
+        return (0);
     return (1);
 }
